+++ CHANGELOG.rst
@@ -6,8 +6,38 @@ All notable changes to `LenslessPiCam
 
 The format is based on `Keep a Changelog <http://keepachangelog.com/en/1.0.0/>`__.
 
+
 Unreleased
 ----------
+
+Added
+~~~~~
+
+- Sensor module.
+
+
+Changed
+~~~~~~~
+
+- 
+
+
+Bugfix
+~~~~~~
+
+- 
+
+1.0.4 - (2023-06-14)
+--------------------
+
+Bugfix
+~~~~~~
+
+- Fix rendering of README on PyPI.
+
+
+1.0.3 - (2023-06-14)
+--------------------
 
 Added
 ~~~~~